Autofill: Extract AutofillParser into separate subproject
