+++ app/build.gradle.kts
@@ -89,7 +89,6 @@ android {
 dependencies {
     compileOnly(Dependencies.AndroidX.annotation)
     implementation(Dependencies.AndroidX.activity_ktx)
-    implementation(Dependencies.AndroidX.autofill)
     implementation(Dependencies.AndroidX.appcompat)
     implementation(Dependencies.AndroidX.biometric)
     implementation(Dependencies.AndroidX.constraint_layout)
@@ -109,6 +108,7 @@ dependencies {
     implementation(Dependencies.Kotlin.Coroutines.android)
     implementation(Dependencies.Kotlin.Coroutines.core)
 
+    implementation(project(Dependencies.FirstParty.autofill_parser))
     implementation(Dependencies.FirstParty.openpgp_ktx)
     implementation(Dependencies.FirstParty.zxing_android_embedded)
