+++ autofill-parser/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest package="com.github.androidpasswordstore.autofillparser">
+</manifest>
