+++ settings.gradle.kts
@@ -2,4 +2,5 @@
  * Copyright © 2014-2020 The Android Password Store Authors. All Rights Reserved.
  * SPDX-License-Identifier: GPL-3.0-only
  */
+include(":autofill-parser")
 include(":app")
